+++ server/src/http/controllers/jobs/jobs.controller.ts
@@ -12,7 +12,7 @@ import dayjs from "../../../common/dayjs.js"
 import { getAppellationDetailsFromAPI, getRomeDetailsFromAPI } from "../../../services/rome.service.js"
 import { getOffre } from "../../../services/formulaire.service.js"
 import { getNearEtablissementsFromRomes } from "../../../services/catalogue.service.js"
-import { ENTREPRISE, etat_utilisateur, validation_utilisateur } from "../../../services/constants.js"
+import { ENTREPRISE, etat_utilisateur, validation_utilisateur } from "../../../services/constant.service.js"
 import { delay } from "../../../common/utils/asyncUtils.js"
 import { ICredential } from "../../../common/model/schema/credentials/credential.types.js"
 import { IApiError } from "../../../common/utils/errorManager.js"
