+++ server/src/http/routes/auth/login.controller.ts
@@ -5,7 +5,7 @@ import passport from "passport"
 import { ExtractJwt, Strategy } from "passport-jwt"
 import { Strategy as LocalStrategy } from "passport-local"
 import { mailTemplate } from "../../../assets/index.js"
-import { CFA, ENTREPRISE, etat_utilisateur } from "../../../services/constants.js"
+import { CFA, ENTREPRISE, etat_utilisateur } from "../../../services/constant.service.js"
 import { UserRecruteur } from "../../../common/model/index.js"
 import { createMagicLinkToken, createUserRecruteurToken, createUserToken } from "../../../common/utils/jwtUtils.js"
 import config from "../../../config.js"
