+++ server/src/http/routes/etablissementRecruteur.controller.ts
@@ -2,7 +2,7 @@ import express, { Request } from "express"
 import joi from "joi"
 import { createFormulaire, getFormulaire } from "../../services/formulaire.service.js"
 import { mailTemplate } from "../../assets/index.js"
-import { CFA, ENTREPRISE, etat_utilisateur, OPCOS, validation_utilisateur } from "../../services/constants.js"
+import { CFA, ENTREPRISE, etat_utilisateur, OPCOS, validation_utilisateur } from "../../services/constant.service.js"
 import { createMagicLinkToken, createUserRecruteurToken } from "../../common/utils/jwtUtils.js"
 import { checkIfUserEmailIsPrivate, checkIfUserMailExistInReferentiel, getAllDomainsFromEmailList } from "../../common/utils/mailUtils.js"
 import { notifyToSlack } from "../../common/utils/slackUtils.js"
